refactor(releases): rename folder from curate to releases
