+++ cernopendata/modules/theme/assets/semantic-ui/js/releases/CurateApp.js
