+++ cernopendata/modules/theme/assets/semantic-ui/js/releases/ReleasesTable.js
