+++ cernopendata/modules/theme/assets/semantic-ui/js/releases/ValidationToggle.js
