+++ cernopendata/modules/theme/assets/semantic-ui/js/releases/documents/EditImageModal.js
