+++ cernopendata/modules/theme/assets/semantic-ui/js/releases/documents/UploadImagesModal.js
