+++ cernopendata/modules/theme/assets/semantic-ui/js/releases/records/BulkEditModal.js
