+++ cernopendata/modules/theme/assets/semantic-ui/js/releases/records/visibility.js
