+++ cernopendata/modules/theme/assets/semantic-ui/js/releases/shared/RowActions.js
